Square Matrix
Write a program to take as input a square matrix, A, and compute the following:
i) |A|
ii) c*A , where c is a scalar value given as user input
iii) The transpose of A
iv) The square of A i.e. A*A
v) The inverse of A
Note that A can be of any dimension n x n and will be given as input via a text file. You can consider the attached file as a sample input.